pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bastion-ssh"
version = "0.1.0"
description = "Dynamic SSH bastion using identity-based routing"
authors = [{name = "Isaac Allard", email = "isaac.allard@epitech.eu"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "paramiko>=3.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

[project.scripts]
bastion = "bastion.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["bastion*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"